chore: fixture 파일 패키지 변경
